+++ exportplan/data.py
@@ -18,3 +18,10 @@ SECTION_TITLES = [
 SECTION_SLUGS = [slugify(section) for section in SECTION_TITLES]
 
 SECTION_URLS = [reverse_lazy('exportplan:section', kwargs={'slug': slug}) for slug in SECTION_SLUGS]
+
+SECTION_TITLES_URLS = [
+    {
+        'title': section,
+        'url': reverse_lazy('exportplan:section', kwargs={'slug': slugify(section)})
+    } for section in SECTION_TITLES
+]