+++ react-components/src/bundle.js
@@ -15,11 +15,13 @@ import CountryFinder from '@src/components/ProductFinder/CountryFinder'
 import { createRouteToMarket, createSpendingAndResources } from '@src/views/sections/MarketingApproach'
 import { aboutYourBusinessForm } from '@src/views/sections/AboutYourBusiness'
 import { adaptToTargetMarketForm } from '@src/views/sections/AdaptationForYourTargetMarket'
+import { sectionSidebar } from '@src/views/sections'
 import { createMarkLessonAsComplete } from '@src/components/MarkLessonAsComplete/MarkLessonAsComplete'
 import { createTargetAgeGroupInsights } from '@src/components/TargetAgeGroupInsights/TargetAgeGroupInsights'
 import { createTargetMarketCountries } from '@src/components/TargetMarketCountries'
 import { createBusinessRationale } from '@src/components/BusinessRationale'
 import { createObjectivesList } from '@src/components/ObjectivesList'
+import { createVideoTranscript } from '@src/components/VideoTranscript/VideoTranscript'
 import LearnIntroduction from '@src/views/LearnIntroduction/LearnIntroduction'
 import { STEP_CREDENTIALS, STEP_VERIFICATION_CODE } from '@src/views/SignupModal/Component/'
 import Services from '@src/Services'
@@ -37,6 +39,7 @@ export default {
   createTargetMarketCountries,
   createBusinessRationale,
   createObjectivesList,
+  createVideoTranscript,
   IndustriesModal,
   LearnIntroduction,
   ProductLookup,
@@ -51,5 +54,6 @@ export default {
   SignupModal,
   STEP_CREDENTIALS,
   STEP_VERIFICATION_CODE,
-  Tour
+  Tour,
+  sectionSidebar
 }