+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - no ticket - Dependencies upgrade
 
 ### Fixed bugs
+
 - No ticket - Fix heading styling
 - GP2-495 - Update headings on lesson pages
 - GP2-514 - Fix alt text
@@ -53,6 +54,9 @@
 - GP2-315 - route to market backend integration
 
 ### Implemented enhancements
+- GP2-263 - Next lesson (BE)
+- GP2-260 - Lesson - Choose don't choose
+- GP2-449 - video transcript FE
 - GP2-594, GP2-595 - Updated personalization bar buttons
 - GP2-449 - Extended video model for transcript field
 - GP2-388 - Country selection with full country list
@@ -221,3 +225,4 @@
 - GP2-420 - get topic details
 - GP2-540 - EP-make-fields-optional
 - GP2-545 - remove airtable
+- GP2-500 - include video duration in read time for the page